+++ src/enums/ObfuscationTarget.ts
@@ -1,10 +1,10 @@
-import { MakeEnum } from '@gradecam/tsenum';
+import { Utils } from '../utils/Utils';
 
 export const ObfuscationTarget: Readonly<{
     Browser: 'browser';
     BrowserNoEval: 'browser-no-eval';
     Node: 'node';
-}> = MakeEnum({
+}> = Utils.makeEnum({
     Browser: 'browser',
     BrowserNoEval: 'browser-no-eval',
     Node: 'node'