+++ src/enums/generators/identifier-names-generators/IdentifierNamesGenerator.ts
@@ -1,11 +1,11 @@
-import { MakeEnum } from '@gradecam/tsenum';
+import { Utils } from '../../../utils/Utils';
 
 export const IdentifierNamesGenerator: Readonly<{
     DictionaryIdentifierNamesGenerator: 'dictionary';
     HexadecimalIdentifierNamesGenerator: 'hexadecimal';
     MangledIdentifierNamesGenerator: 'mangled';
     MangledShuffledIdentifierNamesGenerator: 'mangled-shuffled';
-}> = MakeEnum({
+}> = Utils.makeEnum({
     DictionaryIdentifierNamesGenerator: 'dictionary',
     HexadecimalIdentifierNamesGenerator: 'hexadecimal',
     MangledIdentifierNamesGenerator: 'mangled',