+++ src/enums/node-transformers/string-array-transformers/StringArrayEncoding.ts
@@ -1,10 +1,10 @@
-import { MakeEnum } from '@gradecam/tsenum';
+import { Utils } from '../../../utils/Utils';
 
 export const StringArrayEncoding: Readonly<{
     None: 'none';
     Base64: 'base64';
     Rc4: 'rc4';
-}> = MakeEnum({
+}> = Utils.makeEnum({
     None: 'none',
     Base64: 'base64',
     Rc4: 'rc4'