+++ src/enums/options/presets/OptionsPreset.ts
@@ -1,11 +1,11 @@
-import { MakeEnum } from '@gradecam/tsenum';
+import { Utils } from '../../../utils/Utils';
 
 export const OptionsPreset: Readonly<{
     Default: 'default';
     LowObfuscation: 'low-obfuscation';
     MediumObfuscation: 'medium-obfuscation';
     HighObfuscation: 'high-obfuscation';
-}> = MakeEnum({
+}> = Utils.makeEnum({
     Default: 'default',
     LowObfuscation: 'low-obfuscation',
     MediumObfuscation: 'medium-obfuscation',