+++ src/types/options/TOptionsPreset.ts
@@ -1,5 +1,5 @@
-import { TypeFromEnum } from '@gradecam/tsenum';
+import { TTypeFromEnum } from '../utils/TTypeFromEnum';
 
 import { OptionsPreset } from '../../enums/options/presets/OptionsPreset';
 
-export type TOptionsPreset = TypeFromEnum<typeof OptionsPreset>;
+export type TOptionsPreset = TTypeFromEnum<typeof OptionsPreset>;