+++ src/types/options/TStringArrayEncoding.ts
@@ -1,5 +1,5 @@
-import { TypeFromEnum } from '@gradecam/tsenum';
+import { TTypeFromEnum } from '../utils/TTypeFromEnum';
 
 import { StringArrayEncoding } from '../../enums/node-transformers/string-array-transformers/StringArrayEncoding';
 
-export type TStringArrayEncoding = TypeFromEnum<typeof StringArrayEncoding>;
+export type TStringArrayEncoding = TTypeFromEnum<typeof StringArrayEncoding>;